Separate navbar and header from main content
