+++ src/app/header/header.css
@@ -22,8 +22,8 @@
   border: 5px solid #fff;
   border-radius: 12px;
   box-shadow: 0 1px 1px rgba(136,153,166,0.15);
-  height: 210px;
   position: relative;
+  height: 210px;
   width: 210px;
 }
 .header-avatar-img {
