+++ src/app/main/main.html
@@ -1,5 +1,3 @@
-<app-navbar></app-navbar>
-<app-header></app-header>
 <div class="main-container container">
   <div class="row"></div>
     <div class="sidebar col-md-4">
