+++ src/app/main/main.spec.js
@@ -5,15 +5,15 @@ import {main} from './main';
 describe('main component', () => {
   beforeEach(() => {
     angular
-      .module('app', ['app/main.html'])
+      .module('app', ['app/main/main.html'])
       .component('app', main);
     angular.mock.module('app');
   });
 
-  it('should render the header and techs', angular.mock.inject(($rootScope, $compile) => {
+  it('should render the sidebar and techs', angular.mock.inject(($rootScope, $compile) => {
     const element = $compile('<app>Loading...</app>')($rootScope);
     $rootScope.$digest();
-    expect(element.find('app-header').length).toEqual(1);
+    expect(element.find('app-sidebar').length).toEqual(1);
     expect(element.find('fountain-techs').length).toEqual(1);
   }));
 });
