+++ src/app/main/sidebar/sidebar.js
