+++ src/app/main/sidebar/sidebar.spec.js
