+++ src/index.html
@@ -9,6 +9,9 @@
   </head>
 
   <body ng-app="app">
+    <app-navbar></app-navbar>
+    <app-header></app-header>
+
     <ui-view></ui-view>
   </body>
 </html>
